User-Service/requirements.txt
Flask==3.0.0
flask-cors==3.0.10
psycopg2-binary==2.9.7
kafka-python==2.0.2
supertokens-python==0.12.3
uvicorn==0.23.2
